chore(clean-types): splits out the rest of the types
