+++ runem/cli/initialise_options.py
@@ -2,7 +2,7 @@ import argparse
 
 from runem.config_metadata import ConfigMetadata
 from runem.informative_dict import InformativeDict
-from runem.types_wip.types import OptionsWritable
+from runem.types_wip.options import OptionsWritable
 
 
 def initialise_options(
