+++ runem/types_public_api.py
@@ -1,5 +1,5 @@
 from runem.types_wip.common import FilePathList, JobName
-from runem.types_wip.types import Options
+from runem.types_wip.options import Options
 from runem.types_wip.types_jobs import JobKwargs, JobReturnData
 
 __all__ = ["FilePathList", "JobName", "JobReturnData", "Options", "JobKwargs"]
