+++ runem/types_wip/options.py
@@ -1,0 +1,7 @@
+from runem.informative_dict import InformativeDict, ReadOnlyInformativeDict
+
+OptionName = str
+OptionValue = bool
+OptionsWritable = InformativeDict[OptionName, OptionValue]
+OptionsReadOnly = ReadOnlyInformativeDict[OptionName, OptionValue]
+Options = OptionsReadOnly
